Refactor common receipt parsing code to its own file
This will also be used by the new Feast app pubsub handler.
